modify user and pilot profile

+++ src/main/webapp/js/navbuttons.js
@@ -1,35 +1,68 @@
 $(async function () {
     const current_user = JSON.parse(sessionStorage.getItem("current_user"));
-	if(current_user == null || current_user == undefined){ // user is not logged in
-		document.getElementById("btn-logout").style.display="none";
-        document.getElementById("btn-delete-flight").style.display="none";
-        document.getElementById("btn-plan-flight").style.display="none";
-        document.getElementById("btn-modify-flight").style.display="none";
-        document.getElementById("btn-become-pilot").style.display="none";
-        document.getElementById("btn-user-profile").style.display="none";
-        document.getElementById("btn-modify-pilot-profile").style.display="none";
-	} else { // user is logged in
-		document.getElementById("btn-signup").style.display="none";  
-		document.getElementById("btn-login").style.display="none";
 
-        const pilot = await getPilot(current_user.userID);
+    const btnDeleteFlight = document.getElementById("btn-delete-flight");
+    const btnPlanFlight = document.getElementById("btn-plan-flight");
+    const btnModifyFlight = document.getElementById("btn-modify-flight");
+    const btnBecomePilote = document.getElementById("btn-become-pilot");
+    const btnProfile = document.getElementById("btn-profile");
+    const btnModifyPiloteProfile = document.getElementById("btn-modify-pilot-profile");
+    const btnSignup = document.getElementById("btn-signup");
+    const btnLogin = document.getElementById("btn-login");
+    const btnLogout = document.getElementById("btn-logout");
+    let pilot = null;
+    
+	if (current_user == null || current_user == undefined){ // user is not logged in
+        hideButton(btnLogout);
+        hideButton(btnDeleteFlight);
+        hideButton(btnPlanFlight);
+        hideButton(btnModifyFlight);
+        hideButton(btnBecomePilote);
+        hideButton(btnProfile);
+        hideButton(btnModifyPiloteProfile);
+		
+	} else { // user is logged in
+		hideButton(btnSignup);
+        hideButton(btnLogin);
+
+        pilot = await getPilot(current_user.userID);
         if(!pilot.userID){
             // if user is not pilot
-            document.getElementById("btn-plan-flight").style.display="none";
-            document.getElementById("btn-modify-flight").style.display="none";
-            document.getElementById("btn-delete-flight").style.display="none";
-            document.getElementById("btn-modify-pilot-profile").style.display="none";
+            hideButton(btnPlanFlight);
+            hideButton(btnModifyFlight);
+            hideButton(btnDeleteFlight);
+            hideButton(btnModifyPiloteProfile);
         }else{
             // if user is pilot
-            document.getElementById("btn-become-pilot").style.display="none";
+            hideButton(btnBecomePilote);
         }        
 	}
 
-    document.getElementById("btn-logout").addEventListener("click", ()=>{
-        sessionStorage.clear();
-        window.location.href = "home.html"
-    });
+    if (btnLogout){
+        btnLogout.addEventListener("click", ()=>{
+            sessionStorage.clear();
+            window.location.href = "home.html"
+        });
+    }
+
+    if (btnProfile){
+        btnProfile.addEventListener("click", () => {
+            if(!pilot.userID){
+                // if user is not pilot
+                window.location.href = "userProfile.html";
+            }else{
+                // if user is pilot
+                window.location.href = "pilotInformations.html?pilotID=" + pilot.userID;
+            } 
+        })
+    }
 })
+
+function hideButton(button) {
+    if (button){ 
+        button.style.display="none"; 
+    }
+}
 
 async function getPilot(userID) {
     let result;
@@ -39,7 +72,7 @@ async function getPilot(userID) {
             url: "/ws/coavionnage/pilots/" + userID,
             type: 'GET',
         }).done((response)=>{
-			console.log(response);
+			// console.log(response);
 		}).catch((error)=>{
 			console.error(error);
 		});
